Refactor code and fix solution saving bug
This commit contains some refactoring. At foremost the application now uses the
PapyrusCompilerWrapper to call the official Papyrus compiler. This improves
testing capabilities and maintainability.
As a side effect of the code refactoring a major bug in the solution saving
process could be fixed. Solutions now should save available and selected
configurations properly.